Many Changes
Should be semi-stable. Probably some conditionals that are off or are
not respecting the state. Explored interrupts but their limitations and
room for error superseded their direct need. If we recognize latency
issues later on, interrupts would replace the polling for photocells
ONLY (as they are the trigger for everything else.)